refactor: ui domain rename module
